Try to get data to the parent container more quickly.
I was running some code that does `yield n` every 2 seconds. If the
inmate gets `kill -9`ed, the yielded data never gets to the parent. If
the inmate gets `kill -INT`ed, the yielded data all shows up in a batch
as the inmate exits.

This is an attempt to make the data show up in the parent more quickly.