Add spec test for configuring ExtraHosts option
This adds a new spec test that will mount a local directory and have the docker
container launched copy the `/etc/hosts` file to that directory. This is to
ensure that volume mounts still work as expected (see #17) and that any custom
@container_options passed in to Hoosegow that use the `:HostConfig` key get
properly configured. In this case, we're testing that the `:ExtraHost` option
properly adds the list of hostname/IP mappings to the containers `/etc/hosts`
file.